+++ src/main/java/insane96mcp/enhancedai/modules/mobs/fisher/FisherMobs.java
@@ -1,8 +1,7 @@
-package insane96mcp.enhancedai.modules.mobs;
+package insane96mcp.enhancedai.modules.mobs.fisher;
 
 import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
-import insane96mcp.enhancedai.modules.zombie.ai.FishingTargetGoal;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;