+++ src/main/java/insane96mcp/enhancedai/modules/mobs/fisher/FishingHookRenderer.java
@@ -1,9 +1,8 @@
-package insane96mcp.enhancedai.modules.zombie.client.renderer.entity;
+package insane96mcp.enhancedai.modules.mobs.fisher;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.math.Axis;
-import insane96mcp.enhancedai.modules.zombie.entity.projectile.FishingHook;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRenderer;