+++ src/main/java/insane96mcp/enhancedai/setup/EAEntities.java
@@ -1,8 +1,8 @@
 package insane96mcp.enhancedai.setup;
 
 import insane96mcp.enhancedai.EnhancedAI;
+import insane96mcp.enhancedai.modules.mobs.fisher.FishingHook;
 import insane96mcp.enhancedai.modules.spider.entity.projectile.ThrownWebEntity;
-import insane96mcp.enhancedai.modules.zombie.entity.projectile.FishingHook;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
 import net.minecraftforge.registries.DeferredRegister;